pyproject.toml
[project]
name = "graphrag-agent-runtime"
version = "0.1.0"
description = "Python backend for CopilotKit + LangGraph GraphRAG agent."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "ag-ui-langgraph>=0.0.23",
  "aiosqlite>=0.20.0",
  "copilotkit>=0.1.74",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.0",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pyarrow>=17.0.0",
  "python-dotenv>=1.0.1",
  "uvicorn[standard]>=0.30.0",
]

[tool.uv]
package = false

